pyproject.toml
[tool.poetry]
name = "thothcomm"
version = "1.0.0"
description = "This project aims to allow someone quickly configure Jasmin SMS gateway using a GUI/Web."
authors = ["Emmanuel Keeya <ekeeya@thothcode.tech>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
amqp = "5.2.0"
asgiref = "3.8.1"
attrs = "23.2.0"
automat = "22.10.0"
autopep8 = "2.3.1"
billiard = "4.2.0"
celery = "5.3.6"
certifi = "2024.7.4"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
constantly = "23.10.4"
cryptography = "41.0.7"
django = "5.0.11"
django-filter = "24.2"
djangorestframework = "^3.16.0"
falcon = "3.1.3"
flake8 = "7.1.0"
hyperlink = "21.0.0"
idna = "3.7"
incremental = "22.10.0"
jasmin = "0.11.1"
kombu = "5.3.7"
lockfile = "0.12.2"
markdown = "3.6"
mccabe = "0.7.0"
pexpect = "4.9.0"
prometheus-client = "0.18.0"
prompt-toolkit = "3.0.47"
ptyprocess = "0.7.0"
pyasn1 = "0.5.1"
pyasn1-modules = "0.4.0"
pycodestyle = "2.12.0"
pycparser = "2.22"
pyflakes = "3.2.0"
pyopenssl = "23.3.0"
pyparsing = "3.1.2"
python-dateutil = "2.8.2"
python-messaging = "0.5.13"
python-mimeparse = "1.6.0"
redis = "5.0.7"
requests = "2.31.0"
service-identity = "23.1.0"
six = "1.16.0"
smpp-pdu3 = "0.6"
smpp-twisted3 = "0.8"
sqlparse = "0.5.0"
tabulate = "0.8.10"
treq = "23.11.0"
twisted = "23.10.0"
txamqp3 = "0.9.4"
txredisapi = "1.4.10"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.2"
vine = "5.1.0"
wcwidth = "0.2.13"
xlrd = "2.0.1"
xlwt = "1.3.0"
zope-interface = "6.4.post2"
coverage = "^7.6.0"
psycopg2 = "^2.9.10"
smartmin = "^5.1.2"
pandas = "^2.2.3"
django-timezone-field = "^7.1"
django-countries = {version = "^7.6.1", extras = ["pyuca"]}
iso8601 = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 package-mode = false